+++ src/Inventory/Asset/VirtualMachine.php
@@ -38,7 +38,7 @@ namespace Glpi\Inventory\Asset;
 
 use AutoUpdateSystem;
 use Computer;
-use ComputerVirtualMachine;
+use ItemVirtualMachine;
 use Glpi\Inventory\Conf;
 use RuleImportAssetCollection;
 use Toolbox;
@@ -59,6 +59,9 @@ class VirtualMachine extends InventoryAsset
 
     public function prepare(): array
     {
+        /** @var array $CFG_GLPI */
+        global $CFG_GLPI;
+
         $mapping = [
             'memory'      => 'ram',
             'vmtype'      => 'virtualmachinetypes_id',
@@ -78,8 +81,13 @@ class VirtualMachine extends InventoryAsset
             'macaddr'     => 'mac'
         ];
 
-        if ($this->item->getType() != 'Computer') {
-            throw new \RuntimeException('Virtual machines are handled for computers only.');
+        if (!in_array($this->item->getType(), $CFG_GLPI['itemvirtualmachines_types'])) {
+            throw new \RuntimeException(
+                sprintf(
+                    'Virtual machines are not handled for %s.',
+                    $this->item->getType()
+                )
+            );
         }
 
         foreach ($this->data as &$val) {
@@ -195,9 +203,10 @@ class VirtualMachine extends InventoryAsset
 
         $iterator = $DB->request([
             'SELECT' => ['id', 'name', 'uuid', 'virtualmachinesystems_id'],
-            'FROM'   => ComputerVirtualMachine::getTable(),
+            'FROM'   => ItemVirtualMachine::getTable(),
             'WHERE'  => [
-                'computers_id' => $this->item->fields['id'],
+                'itemtype' => $this->item->getType(),
+                'items_id' => $this->item->fields['id'],
                 'is_dynamic'   => 1
             ]
         ]);
@@ -214,18 +223,18 @@ class VirtualMachine extends InventoryAsset
     public function handle()
     {
         $value = $this->data;
-        $computerVirtualmachine = new ComputerVirtualMachine();
+        $itemVirtualmachine = new ItemVirtualMachine();
         $computer = new Computer();
 
         $db_vms = $this->getExisting();
 
         foreach ($db_vms as $keydb => $arraydb) {
-            $computerVirtualmachine->getFromDB($keydb);
+            $itemVirtualmachine->getFromDB($keydb);
             foreach ($value as $key => $val) {
-                $handled_input = $this->handleInput($val, $computerVirtualmachine);
+                $handled_input = $this->handleInput($val, $itemVirtualmachine);
 
-                //search ComputervirtualMachine on cleaned UUID if it changed
-                foreach (ComputerVirtualMachine::getUUIDRestrictCriteria($handled_input['uuid'] ?? '') as $cleaned_uuid) {
+                //search ItemvirtualMachine on cleaned UUID if it changed
+                foreach (ItemVirtualMachine::getUUIDRestrictCriteria($handled_input['uuid'] ?? '') as $cleaned_uuid) {
                     $sinput = [
                         'name'                     => $handled_input['name'] ?? '',
                         'uuid'                     => $cleaned_uuid ?? '',
@@ -248,7 +257,7 @@ class VirtualMachine extends InventoryAsset
                             }
                         }
 
-                        $computerVirtualmachine->update($input);
+                        $itemVirtualmachine->update($input);
                         unset($value[$key]);
                         unset($db_vms[$keydb]);
                         break 2;
@@ -260,16 +269,17 @@ class VirtualMachine extends InventoryAsset
         if ((!$this->main_asset || !$this->main_asset->isPartial()) && count($db_vms) != 0) {
             // Delete virtual machines links in DB
             foreach ($db_vms as $idtmp => $data) {
-                $computerVirtualmachine->delete(['id' => $idtmp], 1);
+                $itemVirtualmachine->delete(['id' => $idtmp], 1);
             }
         }
 
         if (count($value) != 0) {
             foreach ($value as $val) {
-                $input = $this->handleInput($val, $computerVirtualmachine);
-                $input['computers_id'] = $this->item->fields['id'];
+                $input = $this->handleInput($val, $itemVirtualmachine);
+                $input['itemtype'] = $this->item->getType();
+                $input['items_id'] = $this->item->fields['id'];
                 $input['is_dynamic']  = 1;
-                $computerVirtualmachine->add($input);
+                $itemVirtualmachine->add($input);
             }
         }
 
@@ -384,7 +394,7 @@ class VirtualMachine extends InventoryAsset
             'FROM'   => 'glpi_computers',
             'WHERE'  => [
                 'RAW' => [
-                    'LOWER(uuid)'  => ComputerVirtualMachine::getUUIDRestrictCriteria($vm->uuid)
+                    'LOWER(uuid)'  => ItemVirtualMachine::getUUIDRestrictCriteria($vm->uuid)
                 ]
             ],
             'LIMIT'  => 1
@@ -405,6 +415,6 @@ class VirtualMachine extends InventoryAsset
 
     public function getItemtype(): string
     {
-        return \ComputerVirtualMachine::class;
+        return \ItemVirtualMachine::class;
     }
 }