+++ tests/functional/Search.php
@@ -933,9 +933,10 @@ class Search extends DbTestCase
         ]);
         $this->integer($computer_id)->isGreaterThan(0);
 
-        $cvm = new \ComputerVirtualMachine();
+        $cvm = new \ItemVirtualMachine();
         $cvm_id = $cvm->add([
-            'computers_id' => $computer_id,
+            'itemtype' => 'Computer',
+            'items_id' => $computer_id,
             'name'         => $fname,
             'vcpu'         => 1,
         ]);
@@ -949,7 +950,8 @@ class Search extends DbTestCase
         ]);
         $this->integer($computer_id)->isGreaterThan(0);
         $cvm_id = $cvm->add([
-            'computers_id' => $computer_id,
+            'itemtype' => 'Computer',
+            'items_id' => $computer_id,
             'name'         => $fname,
         ]);
         $this->integer($cvm_id)->isGreaterThan(0);